Partial fix for severe deadlocks
One cannot simply wait for a subprocess to end, then process any output generated, because the buffers involved are small and a deadlock occurs when they become full. This partial fix ensures large *outputs* are accepted by reading all of the output before waiting for the process to end. Large *errors* still cause deadlocks. The only serious solution is threading, but we must investigate Unity support before implementing that.